+++ R/data_edit.R
@@ -290,19 +290,24 @@ data_edit <- function(x = NULL,
     }
     
     # DATA STORAGE
-    values <- reactiveValues(data = NULL, # original data
-                             data_active = NULL, # displayed data
-                             rows = NULL,
-                             columns = NULL,
-                             cut = FALSE)
+    values <- reactiveValues(
+      data = NULL, # original data
+      data_active = NULL, # displayed data
+      rows = NULL,
+      columns = NULL,
+      cut = FALSE,
+      row_index = NULL
+    )
     
     # DATA INPUT
-    data_input <- dataInputServer("input1",
-                                  data = data,
-                                  read_fun = read_fun,
-                                  read_args = read_args,
-                                  hide = hide,
-                                  envir = envir) # search in parent frame
+    data_input <- dataInputServer(
+      "input1",
+      data = data,
+      read_fun = read_fun,
+      read_args = read_args,
+      hide = hide,
+      envir = envir # search in parent frame
+    ) 
     
     # RESET FILTERS
     observeEvent(data_input(), {
@@ -318,16 +323,20 @@ data_edit <- function(x = NULL,
     # FILTERS ALWAYS RESET ON DATA SYNC
     
     # DATA SELECT
-    data_select <- dataSelectServer("select1",
-                                    data = reactive(values$data),
-                                    hide = hide,
-                                    hover_text = "select columns")
+    data_select <- dataSelectServer(
+      "select1",
+      data = reactive(values$data),
+      hide = hide,
+      hover_text = "select columns"
+    )
     
     # DATA FILTER
-    data_filter <- dataFilterServer("filter1",
-                                    data = reactive(values$data),
-                                    hide = hide,
-                                    hover_text = "filter rows")
+    data_filter <- dataFilterServer(
+      "filter1",
+      data = reactive(values$data),
+      hide = hide,
+      hover_text = "filter rows"
+    )
     
     # UPDATE FILTERS
     observe({
@@ -361,19 +370,27 @@ data_edit <- function(x = NULL,
       }
     })
     
+    # ROW INDEX - ROWS IN MASTER COPY
+    observe({
+      values$row_index <- nrow(values$data)
+    })
+    
     # DATAEDIT - ENTIRE DATASET
-    data_update <- dataEditServer("edit1",
-                                  data = reactive({values$data_active}),
-                                  col_bind = NULL, # endless loop!
-                                  col_edit = col_edit,
-                                  col_options = col_options,
-                                  col_stretch = col_stretch,
-                                  col_names = col_names,
-                                  col_readonly = col_readonly,
-                                  col_factor = col_factor,
-                                  row_bind = NULL, # endless loop!
-                                  row_edit = row_edit,
-                                  quiet = quiet)
+    data_update <- dataEditServer(
+      "edit1",
+      data = reactive({values$data_active}),
+      col_bind = NULL, # endless loop!
+      col_edit = col_edit,
+      col_options = col_options,
+      col_stretch = col_stretch,
+      col_names = col_names,
+      col_readonly = col_readonly,
+      col_factor = col_factor,
+      row_bind = NULL, # endless loop!
+      row_edit = row_edit,
+      row_index = reactive({values$row_index}), # row_index + 1 for new rows
+      quiet = quiet
+    )
     
     # UPDATE ACTIVE DATA
     observe({
@@ -381,34 +398,42 @@ data_edit <- function(x = NULL,
     })
     
     # SYNC
-    data_sync <- dataSyncServer("sync1",
-                                data = reactive(values$data),
-                                data_subset = reactive(values$data_active),
-                                rows = reactive(values$rows),
-                                columns = reactive(values$cols),
-                                hide = hide,
-                                hover_text = "synchronise")
+    data_sync <- dataSyncServer(
+      "sync1",
+      data = reactive(values$data),
+      data_subset = reactive(values$data_active),
+      rows = reactive(values$rows),
+      columns = reactive(values$cols),
+      hide = hide,
+      hover_text = "synchronise"
+    )
+    
+    # DATASYNC - ONLY UPDATE MASTER - REMOVE FILTERS FOR DISPLAY
     observe({
       values$data <- data_sync()
     })
     
     # DATA OUTPUT - DATA ACTIVE
-    dataOutputServer("output-active",
-                     data = reactive({values$data_active}),
-                     save_as = save_as,
-                     write_fun = write_fun,
-                     write_args = write_args,
-                     hide = hide,
-                     hover_text = "save selection \n to file")
+    dataOutputServer(
+      "output-active",
+      data = reactive({values$data_active}),
+      save_as = save_as,
+      write_fun = write_fun,
+      write_args = write_args,
+      hide = hide,
+      hover_text = "save selection \n to file"
+    )
     
     # DATA OUTPUT - DATA ENTIRE
-    dataOutputServer("output-update",
-                     data = reactive({values$data}),
-                     save_as = save_as,
-                     write_fun = write_fun,
-                     write_args = write_args,
-                     hide = hide,
-                     hover_text = "save to file")
+    dataOutputServer(
+      "output-update",
+      data = reactive({values$data}),
+      save_as = save_as,
+      write_fun = write_fun,
+      write_args = write_args,
+      hide = hide,
+      hover_text = "save to file"
+    )
     
     # CUT
     observeEvent(input$cut, {